# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MemoryMesh API — AI Meeting Memory System powered by Cognee
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/topics:
    get:
      operationId: getTopTopics
      tags: [dashboard]
      summary: Get top discussed topics
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicFrequency"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── TIMELINE ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TopicFrequency:
      type: object
      required: [topic, count, last_mentioned]
      properties:
        topic:
          type: string
        count:
          type: integer
        last_mentioned:
          type: string
          format: date-time
        trend:
          type: string
          enum: [rising, stable, declining]
    ActivityItem:
      type: object
      required: [id, type, title, timestamp]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [meeting_added, decision_made, task_created, decision_updated, entity_merged, memory_improved]
        title:
          type: string
        description:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
        entity_id:
          type: ["integer", "null"]
        meeting_id:
          type: ["integer", "null"]

    # ─── TIMELINE ──────────────────────────────────────────────────────────────
...